+++ scripts/settings/dnd5e-settings.js
@@ -6,7 +6,7 @@ export function registerSettings(app, updateSettings) {
         scope: "client",
         config: true,
         type: Boolean,
-        default: true,
+        default: false,
         onChange: value => { updateSettings(value); }
     });
 
@@ -36,7 +36,7 @@ export function registerSettings(app, updateSettings) {
         scope: "client",
         config: true,
         type: Boolean,
-        default: true,
+        default: false,
         onChange: value => { updateSettings(value); }
     });
 
@@ -46,7 +46,7 @@ export function registerSettings(app, updateSettings) {
         scope: "client",
         config: true,
         type: Boolean,
-        default: true,
+        default: false,
         onChange: value => { updateSettings(value); }
     });